Extract component for ‘Melding’ section
