+++ src/app/features/cases/components/organisms/CaseDetail/CaseDetailSectionSignal.tsx
@@ -1,0 +1,74 @@
+import React, { FC } from "react"
+import styled from "styled-components"
+
+import { useCase } from "app/state/rest/index"
+import formatDate from "app/features/shared/utils/formatDate"
+import replaceNewLines from "app/features/shared/utils/replaceNewLines"
+import replaceUrls from "app/features/shared/utils/replaceUrls"
+import highlightText from "app/features/shared/utils/highlightText"
+import Purified from "app/features/shared/components/molecules/Purified/Purified"
+import { KeyValueDetail } from "app/features/types"
+import CaseDetailSection from "./CaseDetailSection"
+import { HrSpaced } from "./CaseDetailSectionStyles"
+
+type Props = {
+  caseId: string
+}
+
+const P = styled.p`
+  margin: 0 0 8px;
+`
+
+const CaseDetailSectionSignal: FC<Props> = ({ caseId }) => {
+  const { data: caseData } = useCase(caseId)
+
+  if (!caseData?.bwv_hotline_melding.length) {
+    return null
+  }
+
+  const meldingen = caseData?.bwv_hotline_melding.map(melding => {
+    const {
+      melding_datum: datum,
+      melding_anoniem: anoniem,
+      melder_naam: naam,
+      melder_telnr: telnr,
+      situatie_schets: text
+    } = melding
+
+    return {
+      datum: datum ? formatDate(datum, true)! : undefined,
+      anoniem: anoniem === "J",
+      naam,
+      telnr,
+      text: replaceNewLines(replaceUrls((text || "").trim(), "_blank"))
+    }
+  }).reverse()
+
+  const meldingenData = meldingen.reduce((acc, item, index) => {
+    const { datum, anoniem, naam, telnr, text } = item
+    const highlightedText = highlightText([ "hoofdhuurder", "hoofdhuur", "hh" ], text, { caseSensitive: false })
+
+    acc.push([ "Datum melding", datum || "-" ])
+    acc.push([ "Anonieme melding", anoniem ])
+    acc.push([ "Melder", <P className="anonymous"> { naam }</P> || "-" ])
+    acc.push([ "Melder telefoonnummer", telnr ?
+      <a className="anonymous" href={ "tel://" + telnr }>{ telnr }</a> : "-" ])
+    acc.push(<Purified className="anonymous" text={ highlightedText } />)
+
+    if (index < meldingen.length - 1) {
+      acc.push(<HrSpaced />)
+    }
+
+    return acc
+  }, [] as KeyValueDetail[])
+
+  return (
+    <CaseDetailSection
+      title="Meldingen / aanleiding"
+      dataSource="BWV"
+      data={ meldingenData.length ? meldingenData : [ "-" ] }
+    />
+  )
+}
+
+export default CaseDetailSectionSignal
