+++ src/app/features/cases/components/organisms/CaseDetail/CaseDetailSectionStyles.tsx
@@ -1,8 +1,13 @@
 import styled from "styled-components"
+import Hr from "app/features/cases/components/atoms/Hr/Hr"
 
 export const CenteredAnchor = styled.a`
   display: block;
   text-align: center;
+`
+
+export const HrSpaced = styled(Hr)`
+  margin: 24px 0;
 `
 
 export const Section = styled.section`
